ignore .python-version in docker build
this file is used by pyenv to set the python version to be used for the current directory. The docker files copied to file and would then throw an error the specific python version is not installed. The docker images provide their own python version, so the file should not be copied into the container

Signed-off-by: Timo Glastra <timo@animo.id>