fix: Add sent key on creddef & schema to standardize response
The OpenAPI Spec schema indicates that "sent" is a possible field,
however, no code paths (aside from the changes in the last two commits)
returned it. Seeing as the last two commits return it during an error
case, the changes here now return the same data during a successful
case. This should make it easier for controllers to grab the `schema_id`
or the `cred_def_id` under both circumstances.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>