fix: Standardize SchemaID/CredDefID responses for non-authors
Add the `sent` key to the response for non-authors to follow the OpenAPI
Schema Specification. Due to the fact that removing the old values would
be a breaking change, I have left those values in for the time being to
allow for a migration path to the new `sent` key.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>