fix: Allow ability to use credentialStatus in JSONLD Creds
Following the documentation for JSONLD Credentials, there is a parameter
for adding `credentialStatus` to a credential:
https://github.com/hyperledger/aries-rfcs/blob/main/features/0593-json-ld-cred-attach/README.md

However, adding the parameter causes all flows to fail due to the
credentialStatus never making it into the credential. I don't know the
history of this parameter, but in order to follow RFC 0593 a bit more
closely, I have fixed up the flow so that it doesn't fail issuance.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>