fix: Safely shutdown when root_profile uninitialized
If an exception was thrown during the initialization of the
root_profile, we would throw an exception during shutdown and fail to
end the process. This would leave ACA-Py running, without
accepting/handling any further requests. Adding guards for shutdown code
that uses the root_profile should avoid this situation

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>