fix: ensure request matches offer, if sent
This change implements checking that ld proof credential requests match
their corresponding offer, if an offer was sent.

If using the `--auto-*` flags for issuance, it was possible for the
receiver of the credential offer to change values in the request and
ACA-Py would accept this and issue based off of the request values. The
`--auto-*` flags are debug flags and should not be used in production
which would mean that a controller should have been able to catch this
discrepancy. However, it is still expedient for ACA-Py to check that the
offer and request match to avoid this slipping past the controller as
well.

There is a side affect of this check. We were permitting late binding of
the credential subject ID to the holder in the request. Meaning, on
request, the holder will automatically (when using auto flags) insert a
DID key as the credential subject ID to ensure the holder can actually
present proof of possession later. These changes modify this behavior
such that it only applies iff the credential subject id is not set
already (e.g. in the credential offer). This enables the issuer to bind
the credential to a DID other than the holder's pairwise DID if an
alternate is known to the issuer. If the issuer wants to permit late
binding by the holder still, the credential subject ID should be omitted
in the offer.

So, to summarize, the two modifications implemented here:
- Ensure the request doesn't change the credential unless the offer
  explicitly omits a credential subject ID
- Only override with holder did if the credential subject ID is omitted

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>