fix: ignore duplicate record errors on add key
This is a bandaid solution to prevent routing keys from being stored for
multiple DIDs. It would be nice to correct behavior of the DIDDoc class
to not turn routing keys into values in the publicKey list and then
claiming that the DID is the controller (with no evidence). However,
given the impending replacement of this behavior with did:peer, it
doesn't seem worth it to go through that effort right now.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>