fix: version should be set by pyproject.toml
`aries_cloudagent.version.__version__` was out of sync with the version
listed in the pyproject.toml file. The pyproject.toml file should be the
source of truth as it will be used when publishing using poetry.
However, `__version__` is used in code for some operations. To ensure
that the pyproject file and the `__version__` can't get out of sync,
`aries_cloudagent.version` will now read from the package metadata to
obtain the version.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>