refactor: replace multiformats library
This replaces the multiformats library with a very basic implementation
included directly within ACA-Py. Given that the multiformats library has
gone stale and my PR has languished, I think this is a good alternative
to using that library. The implementation is very simple and currently
only supports the minimum currently required by ACA-Py. Expanding the
implementation in the future should be trivial.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>