fix: clean up requests and invites
On connection deletion, all records strongly associated with the
connection should be deleted. These records include connection metadata,
connection requests, and connection invitations.

Records referencing the connection ID such as mediation or routing
records are not "strongly" associated with the connection; there are
other mechanisms in place to enable deletion of these records through
the Admin API. The records for metadata, requests, and invitations, on
the other hand, cannot be cleaned up any other way.

This PR makes sure these records are deleted when the connection is
deleted.

It might be a good idea for us to look at other records that ought to be
cleaned up automatically. Those can be addressed in future work.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>