fix: save multi_use to the DB for OOB invitations
It was reported by a collegue that when using Out-of-band invitations
with the connections handshake protocol, along with a public DID and
connection re-use enabled, ACA-Py was only able to handle a single
invitations. As if Connection Re-use wasn't working or the multi_use
flag was ignored.

Turns out, we flat out weren't saving the multi_use flag to the database
which is a big problem if you want to re-use the invitation. When
DID-Exchanged is used, since DID-Exchange allows for a connection-less
exchange of information. We only say the behaviour of the OOB Invitation
getting deleted if we ran through the receive_reuse_message() method in
the manager, which deletes the record if it's not set for multi_use.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>