fix: print provision messages if auto-provision is triggered
It was recently discovered that if you set a static seed, and you have
auto-provisioning enabled on the start command, there is no way to get
the verkey for the created DID. The verkey is never printed out to the
command line and I do not remember there being an API for retrieving the
verkey (one of those, "See it once, copy it, or forever lose it" kinds
of things).

After this point in the code, the only thing that "provision" enables
are the print statements talking about the did/verkey being created as
well as some basic profile information. It doesn't trigger any code that
would make changes to the system, those lines precede the line I've
added.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>