feat: simplify request handling in prep for more protocols
The complexity of a single protocol for DIDComm V2 was a bit much, and
adding more protocols would cause unnecessary copy-pasting of
boilerplate code that could otherwise be pulled out into their own
function.

For the time being, I'm planning to include all protocols for the
DIDComm V2 Interop within the same protocol plugin (and the same file).
In the future, I plan to break this logic out of this file and into
their own respective files. I also had a discussion with @dbluhm today
about future plans for retrieving our DID for DIDComm V2, as the current
implementation is very fragile and requires that, not only is a DID of
the appropriate type already created, but that we grab the latest one.
For a production system, this is unnacceptable. But for a "cobbled
together POC" for the interop? It should be fine. Especially considering
that we have plans to change how this all works in the future.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>