:construction_worker: Skip sonar-merge-main workflow if github actor is dependabot (#3691)
This is because GitHub Actions does not provide repository secrets (such as SONAR_TOKEN) to workflows triggered by dependabot[bot] for security reasons. As a result, the SonarCloud scan step fails when dependabot merges to main (i.e., using `@dependabot squash and merge`, instead of manually merging oneself). By skipping the scan when the actor is dependabot, we avoid unnecessary workflow failures.

Signed-off-by: ff137 <ff137@proton.me>