fix: update tails server upload methods to return public file URIs
We ran into an issue where we had a different tails_server_base_url from
the tails_server_upload_url. This makes sense where you want to have a
separate base URL for accessing the public files, and a private URL that
is used for pushing the changes (where access controls dictate who can
upload files, and they are restricted on the public URL).

The `upload_tails_file` methods were returning the private URLs and the
calling code was expecting public URLs. The solution was to update the
methods to return the public URLs instead, while still uploading to the
private URLs.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>