feat: Add option to remove credex on failure
Something that we noticed was that credential exchange records (and
subsequently, presentation exchange records) were not being removed upon
failure. Even if the `auto_remove` flag was specified. This results in
the controller having to remove each failed record by hand, even if the
controller received all of the data from the save webhook already.

I believe that it is still valuable for failed records to linger
independent of the successful records, since the failed records contain
logs pertaining to their failure.

So I propose a new API flag, `auto_remove_on_failure`, which removes the
records upon failure. In addition to adding the flag, I have added a CLI
parameter, `--no-preserve-failed-exchange-records`, which does the same
thing.

The reason for prefixing the parameter with `no-` is to prevent breaking
existing behavior if users are currently relying on failed/abandoned
records being preserved.

After this change, the controller will no longer need to delete the
records manually. They will still receive the status update via the
`save` call on the record prior to deletion of the record.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>